added modelPredictionManager
